pyproject.toml
[tool.poetry]
name = "backend-event-manager-app"
version = "0.1.0"
description = "Backend for event management application"
authors = ["Ekaterina"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.0"
pydantic = "^2.7.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.20.0"}
aerich = "^0.7.1"
httpx = "^0.28"  # Основная версия
pydantic-settings = "^2.1.0"
sentry-sdk = "^2.37"
pydantic-i18n = "^0.4.5"
fastapi-babel = "^1.0.0"
fastapi-pagination = "^0.12.7"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pyjwt = "^2.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
typing-extensions = "^4.9.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.4"
pytest = "^7.4.3"
requests = "^2.28.1"
pytest-asyncio = "^0.21.1"
asgi-lifespan = "^2.1.0"
# Убрал httpx отсюда - он уже в основных зависимостях

[tool.aerich]
tortoise_orm = "config.tortoise_settings"
location = "app/db/migrations"
src_folder = "./app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "app",
  ".",
]
addopts = [
    "--import-mode=importlib",
    "--asyncio-mode=auto",
]
#asyncio_default_fixture_loop_scope = "module"